requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
configparser==3.5.0
coverage==4.5.1
coveralls==1.5.1
cryptography==2.3.1
docopt==0.6.2
enum34==1.1.6
fileconfig==0.5.7
Flask-Cors==3.0.6
Flask==1.0.2
Flask-RESTful==0.3.6
futures==3.1.1
gunicorn==19.9.0
idna==2.7
ipaddress==1.0.22
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
psycopg2==2.7.5
pycparser==2.19
PyJWT==1.6.4
pylint==1.9.3
pyOpenSSL==18.0.0
python-dotenv==0.9.1
pytz==2018.5
requests==2.19.1
singledispatch==3.4.0.3
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
